requirements.txt
# Smart Student Attendance System - Dependencies
# Install: pip install -r requirements.txt

# Web Framework
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3

# Database
SQLAlchemy==2.0.36

# Face Recognition (DeepFace ArcFace; Windows-friendly, no dlib)
deepface>=0.0.79

# Image Processing
Pillow==11.1.0
numpy>=1.24.0,<3.0.0
opencv-python>=4.8.0

# QR Code Generation
qrcode[pil]==8.0

# Environment Variables
python-dotenv==1.0.1

# Timezone (IST); only needed on Python 3.8
backports.zoneinfo>=0.2.1; python_version < "3.9"

# WSGI Server (production)
gunicorn==23.0.0; sys_platform != 'win32'
waitress==3.0.2; sys_platform == 'win32'
